requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# DB & settings
sqlalchemy[asyncio]>=2.0.36
aiosqlite>=0.20.0
pydantic-settings>=2.6.0

# HTTP for Groq
httpx>=0.27.0

# Retries
tenacity>=9.0.0

# Dev / test
pytest>=8.3.0
pytest-asyncio>=0.24.0
asgi-lifespan>=2.1.0
